refactor: split py_queue/workflow_ops.rs into submodule
Move 1004-line workflow_ops.rs into a folder module split by concern:
lifecycle, nodes, fan_out, gates, queries. Each submodule holds a
partial #[pymethods] impl block. No behavior changes.
